fix: A bit more support for Cepheus Atom (and friends), there is now a system setting to enable showing only END, Lifeblood and (optionally) Contamination instead of the regular characteristics. (Actually uses STR for Lifeblood and PSI for Contamination).
This is a Dreadful Hack that will undoubtedly Break Things Horribly if looked at too closely. If anyone asks, mutant marauders from the wastelands forced me to do it. :)

I.e. I hacked this together without much (or, well, any) concern for how I had originally intended to do this. If/when this gets redesigned it's probably best to just rip out this commit.