feat: add buttons to component hits list (#905)
This release adds buttons to adjust component hits (when viewed on ship sheet using the single line option).  Also there is a new ruleset setting for the maximum number of hits until a component is destroyed.  This setting sets the ceiling for the hits count.  The default value is three.  When this value is reached when changing hits with the buttons, the status is automatically set to "destroyed".