fix: Missing actor errors with ship positions (#994)
This fix (hopefully) corrects the problem with deleted actors and ship positions as noted by @ZEE.  The system now attempts to remove an actor from any ship positions if it that actor is deleted.  Also, missing actors are not shown on positions. I avoid a migration as I was having issues with it for another pull request.