feat: add melee range modifier (#1519)
This small update adds a melee range modifier for weapons. The modifier is applied when the weapon is within melee combat distance.  There is also a new setting to define that melee combat distance. This new feature is useful when using ranged weapons within melee range.  Note the modifier doesn't apply for weapons with the range set to 'Melee', i.e. melee weapons.  Also, the modifier doesn't apply when using the CE weapon range method - as it is already adjusts for melee range.
Thanks to @ZEE for the suggestion.