feat: Alpha Cephei ruleset and compendia (#1614)
This release adds support and integration for the Alpha Cephei ruleset. It includes compendia for skills, items and actors as well as updates to the automation.  A big thanks to RafaelLVX for permission and hard work in bringing this ruleset to the Twodsix community.  More about Alpha Cephei can be found here: https://alphacepheirpg.com